ADD: ユーザー削除用のAPIを作成

+++ routes/users.js
@@ -23,11 +23,15 @@ router.put('/:id', async (req, res) => {
 
 // Delete a user
 router.delete('/:id', async (req, res) => {
-    try {
-        await User.findByIdAndDelete(req.params.id);
-        res.status(200).json({ message: 'User deleted successfully' });
-    } catch (err) {
-        return res.status(500).json(err);
+    if (req.body.userId === req.params.id || req.body.isAdmin) {
+        try {
+            await User.findByIdAndDelete(req.params.id);
+            res.status(200).json({ message: 'User deleted successfully' });
+        } catch (err) {
+            return res.status(500).json(err);
+        }
+    } else {
+        return res.status(403).json('You can only delete your own account!');
     }
 });
